Simplify the way fragments are parsed
