+++ .travis.yml
@@ -10,10 +10,6 @@ script:
   - mkdir target
   - cd render_tests
   - bash run.sh
-  - cd ../pattern
-  - cargo doc
-  - mkdir ../tokenizer/target
-  - mv target/doc/ ../tokenizer/target/doc
   - cd ../tokenizer
   - cargo test
   - cargo doc
