+++ core/Cargo.toml
@@ -3,14 +3,8 @@ name = "symbiosis_core"
 version = "0.0.1"
 authors = ["Erik Hedvall <hello@erikhedvall.nu>"]
 
-build = "build.rs"
-
 [lib]
 name = "symbiosis_core"
 
 [dependencies]
 tendril = "0.2"
-lalrpop-util = "0.11"
-
-[build-dependencies.lalrpop]
-varsion = "0.11"
