--- core/src/pattern/.gitignore
@@ -1,1 +1,0 @@
-grammar.rs
