--- tokenizer/src/fragment/generated.rs
@@ -1,1 +1,0 @@
-include!(concat!(env!("OUT_DIR"), "/pattern.rs"));
