--- tokenizer/src/parser/.gitignore
@@ -1,1 +1,0 @@
-grammar.rs
